Added react-navigator
