+++ locator-app/App.js
@@ -1,53 +1,16 @@
-import React from 'react';
-import {
-    Text
-  , View
-  , StyleSheet
-  , TextInput
-  , Button
-} from 'react-native';
+import { createStackNavigator, createAppContainer } from 'react-navigation';
+import Homescreen from './Homescreen';
+import Routes from './Rotes';
+
+const AppStack = createStackNavigator({
+  Home: {screen: HomeScreen},
+  Routes:  { screen: Routes }
+});
+
+const AppContainer = createAppContainer(AppStack);
 
 export default class App extends React.Component {
-
   render() {
-    let stub = () => {};
-    return (
-      <View style={styles.container}>
-        <Text style={styles.text}>
-          Enter your bus route number to display buses near you.
-        </Text>
-        <TextInput
-          style={{height: 40, width: 100}}
-          placeholder="Route Number"
-          onChangeText={(text) => this.setState({text})}
-        />
-        <Button
-          onPress={stub}
-          title="N"
-          color="#841584"
-          accessibilityLabel="Learn less about this green button"
-        />
-      </View>
-    );
-
+    return <AppContainer />;
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-    margin: 0,
-
-  },
-  text: {
-    fontSize: 16,
-    fontWeight: '400',
-    //lineHeight: 1.5,
-    color: '#212529',
-    height: 40
-    //textAlign: 'left'
-  }
-});
