+++ locator-app/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
-    "react-native-scripts": "1.14.0",
     "jest-expo": "~27.0.0",
+    "react-native-scripts": "1.14.0",
     "react-test-renderer": "16.3.1"
   },
   "main": "./node_modules/react-native-scripts/build/bin/crna-entry.js",
@@ -21,7 +21,8 @@
   "dependencies": {
     "expo": "^27.0.1",
     "react": "16.3.1",
-    "react-native": "^0.57.3",
-    "react-navigation": "^2.18.0"
+    "react-native": "^0.57.8",
+    "react-native-gesture-handler": "^1.0.12",
+    "react-navigation": "^2.18.3"
   }
 }
